修改返回值

+++ src/main/java/com/thinkgem/jeesite/modules/contract/service/LeaseContractService.java
@@ -134,7 +134,7 @@ public class LeaseContractService extends CrudService<LeaseContractDao, LeaseCon
 			paymentTrans.setUpdateDate(new Date());
 			paymentTrans.setUpdateBy(UserUtils.getUser());
 			paymentTrans.setDelFlag("0");
-			if(0!=deposit)
+			if (0 != deposit)
 				paymentTransDao.insert(paymentTrans);
 
 			// 2.房租款项
@@ -144,7 +144,7 @@ public class LeaseContractService extends CrudService<LeaseContractDao, LeaseCon
 			List<LeaseContractDtl> list = leaseContractDtlDao.findList(leaseContractDtl);
 			for (LeaseContractDtl tmpLeaseContractDtl : list) {
 				// 计算开始日期与结束日期之间的月数
-				float month = DateUtils.getMonthSpace(tmpLeaseContractDtl.getStartDate(),
+				double month = DateUtils.getMonthSpace(tmpLeaseContractDtl.getStartDate(),
 						tmpLeaseContractDtl.getEndDate());
 				month = (month == 0 ? month++ : month);
 				for (int i = 1; i <= month; i++) {
@@ -169,7 +169,7 @@ public class LeaseContractService extends CrudService<LeaseContractDao, LeaseCon
 					paymentTrans.setUpdateDate(new Date());
 					paymentTrans.setUpdateBy(UserUtils.getUser());
 					paymentTrans.setDelFlag("0");
-					if(0!=tmpLeaseContractDtl.getDeposit())
+					if (0 != tmpLeaseContractDtl.getDeposit())
 						paymentTransDao.insert(paymentTrans);
 				}
 			}
